utils.shell: new module for shell utils
moved `which` from `io.kerberos` and added `call`
